fix(core): split executePipeline tests, add timeoutMs coverage, fix dead code
- Extract executePipeline tests into pipeline-executor.test.ts
- Add test for timeoutMs option (PIPELINE_TIMEOUT skipping)
- Add suggestion to PIPELINE_TIMEOUT error per AFD conventions
- Remove unreachable stopped variable and if-guard
